fix(mdc-select): Remove support reliant on native select.
BREAKING CHANGE: material-components-web v0.30.0 removes support for CSS only select element. Note a future release of material-components-web plans to replace underlying mdc-menu with native html select.
