+++ components/select/README.md
@@ -11,17 +11,13 @@
 
 ```javascript
 var vm = new Vue({
-  data () {
+  data() {
     return {
       selected: '',
-    }
-  }
-})
+    };
+  },
+});
 ```
-
-> For better cross-device experiennce, `<mdc-select>` will switch from menu to
-> native rendering based on the media query
-> `(max-width: 600px) and (pointer: coarse)`.
 
 ### mdc-select
 
@@ -30,9 +26,7 @@ var vm = new Vue({
 | `disabled` | Boolean         | false   | whether this select is disabled              |
 | `label`    | String          |         | the selection label                          |
 | `value`    | [String, Array] |         | bind with v-model (update on `change` event) |
-| `multiple` | Boolean         |         | whether the select accept multiple values    |
 | `max-size` | Boolean         | 4       | multi select max size                        |
-| `native`   | String          |         | force native select rendering                |
 | `name`     | String          |         | native select name                           |
 
 | event    | Description             |
@@ -57,30 +51,6 @@ var vm = new Vue({
 </mdc-select>
 ```
 
-### Multiple select
-
-```html
-<mdc-select multiple v-model="selected" label="Please select one or more">
-  <mdc-option>A</mdc-option>
-  <mdc-option>B</mdc-option>
-  <mdc-option>C</mdc-option>
-</mdc-select>
-<span>Selected: {{ selected }}</span>
-```
-
-```javascript
-var vm = new Vue({
-  data () {
-    return {
-      selected: [],
-    }
-  }
-})
-```
-
-> multi-select are always rendered natively. the size is reactive and is caped
-> by the max-size property
-
 ### Reference
 
-- <https://material.io/components/web/catalog/input-controls/select-menus>
+* <https://material.io/components/web/catalog/input-controls/select-menus>
