+++ components/select/demo.vue
@@ -16,41 +16,31 @@
       <mdc-caption tag="p" v-if="selectedType"
       >Selected Value: {{ selectedValue }}</mdc-caption>
       <br>
-
-      <mdc-select multiple v-model="selectedValues" label="Pick one or more"
-      v-if="selectedType">
-      <mdc-option v-for="option of options" :key="option"
-      :value="option.toLowerCase()">
-      {{option}}
-      </mdc-option>
-      </mdc-select>
-      <mdc-caption tag="p" v-if="selectedType"
-      >Multi Select: {{ selectedValues }}</mdc-caption>
     </div>
   </div>
 </template>
 
 <script>
-  export default {
-    data () {
-      return {
-        selectedType: undefined,
-        selectedValue: undefined,
-        selectedValues: [],
-        food: {
-          'Vegetables' : ['Spinach', 'Carrots', 'Onions', 'Broccoli'], 
-          'Meat' : [ 'Eggs', 'Chicken', 'Fish', 'Turkey', 'Pork' , 'Beef'], 
-          'Fruits' : ['Apples', 'Oranges', 'Bananas', 'Berries', 'Lemons'], 
-        }
-      }
+export default {
+  data() {
+    return {
+      selectedType: undefined,
+      selectedValue: undefined,
+      selectedValues: [],
+      food: {
+        Vegetables: ['Spinach', 'Carrots', 'Onions', 'Broccoli'],
+        Meat: ['Eggs', 'Chicken', 'Fish', 'Turkey', 'Pork', 'Beef'],
+        Fruits: ['Apples', 'Oranges', 'Bananas', 'Berries', 'Lemons'],
+      },
+    };
+  },
+  computed: {
+    types() {
+      return Object.keys(this.food);
     },
-    computed: {
-      types () {
-        return Object.keys(this.food)
-      }, 
-      options () {
-        return this.selectedType ? this.food[this.selectedType] : []
-      } 
-    }
-  }
+    options() {
+      return this.selectedType ? this.food[this.selectedType] : [];
+    },
+  },
+};
 </script>
