+++ components/select/mdc-option.vue
@@ -1,43 +1,19 @@
 <template>
-  <component :is="type" :disabled="disabled" :value=value>
-    <slot></slot>
-  </component>
+   <li class="mdc-option mdc-list-item" 
+    role="option" 
+    :tabindex="disabled?-1:0"
+    :aria-disabled="disabled"
+    :data-value="value">
+     <slot></slot>
+   </li>
 </template>
 
 <script>
-import MDCNativeOption from './mdc-native-option.vue';
-import MDCMenuOption from './mdc-menu-option.vue';
-import MDCMultiOption from './mdc-multi-option.vue';
-
 export default {
   name: 'mdc-option',
   props: {
     value: String,
     disabled: Boolean,
-  },
-  inject: ['mdcSelect'],
-  components: {
-    'mdc-native-option': MDCNativeOption,
-    'mdc-multi-option': MDCMultiOption,
-    'mdc-menu-option': MDCMenuOption,
-  },
-  computed: {
-    isNative() {
-      return this.mdcSelect.isNative;
-    },
-    multiple() {
-      return this.mdcSelect.multiple;
-    },
-    menu() {
-      return this.mdcSelect.menu;
-    },
-    type() {
-      return this.multiple
-        ? 'mdc-multi-option'
-        : this.menu
-          ? 'mdc-menu-option'
-          : this.isNative ? 'mdc-native-option' : 'mdc-menu-option';
-    },
   },
 };
 </script>
