+++ components/select/test.spec.js
@@ -1,11 +1,7 @@
 import { mount, createLocalVue, checkValidMdcAdapter } from '../unit-test'
 
-import mdcMenuSelect from './mdc-menu-select.vue'
-import mdcMenuOption from './mdc-menu-option.vue'
-import mdcNativeSelect from './mdc-native-select.vue'
-import mdcNativeOption from './mdc-native-option.vue'
-import mdcMultiSelect from './mdc-multi-select.vue'
-import mdcMultiOption from './mdc-multi-option.vue'
+import mdcSelect from './mdc-select.vue'
+import mdcOption from './mdc-option.vue'
 
 import Spec from './test.spec.vue';
 
@@ -19,27 +15,12 @@ describe(__dirname, () => {
       expect(spec.isVueInstance()).toBeTruthy()
     })
 
-    describe('mdcMenuSelect', () => {
-      checkValidMdcAdapter(spec.find(mdcMenuSelect).vm)
+    describe('mdcSelect', () => {
+      checkValidMdcAdapter(spec.find(mdcSelect).vm)
     })
-    describe('mdcMenuOption', () => {
-      checkValidMdcAdapter(spec.find(mdcMenuOption).vm)
+    describe('mdcOption', () => {
+      checkValidMdcAdapter(spec.find(mdcOption).vm)
     })
-
-    describe('mdcNativeSelect', () => {
-      checkValidMdcAdapter(spec.find(mdcNativeSelect).vm)
-    })
-    describe('mdcNativeOption', () => {
-      checkValidMdcAdapter(spec.find(mdcNativeOption).vm)
-    })
-
-    describe('mdcMultiSelect', () => {
-      checkValidMdcAdapter(spec.find(mdcMultiSelect).vm)
-    })
-    describe('mdcMultiOption', () => {
-      checkValidMdcAdapter(spec.find(mdcMultiOption).vm)
-    })
-
   })
 
 })
