+++ components/select/test.spec.vue
@@ -1,38 +1,20 @@
 <template>
   <div>
-    <mdc-menu-select v-model="selectValue" label="select label">
-      <mdc-menu-option>option1</mdc-menu-option>
-      <mdc-menu-option value="option2">Option2</mdc-menu-option>
-    </mdc-menu-select>
-
-    <mdc-native-select v-model="selectValue" label="select label">
-      <mdc-native-option>option1</mdc-native-option>
-      <mdc-native-option value="option2">Option2</mdc-native-option>
-    </mdc-native-select>
-
-    <mdc-multi-select v-model="selectValues" label="multi select label">
-      <mdc-multi-option>option3</mdc-multi-option>
-      <mdc-multi-option>option3</mdc-multi-option>
-    </mdc-multi-select>
+    <mdc-select v-model="selectValue" label="select label">
+      <mdc-option>option1</mdc-option>
+      <mdc-option value="option2">Option2</mdc-option>
+    </mdc-select>
 
   </div>
 </template>
 
 <script>
-import mdcMenuSelect from './mdc-menu-select.vue'
-import mdcMenuOption from './mdc-menu-option.vue'
-import mdcNativeSelect from './mdc-native-select.vue'
-import mdcNativeOption from './mdc-native-option.vue'
-import mdcMultiSelect from './mdc-multi-select.vue'
-import mdcMultiOption from './mdc-multi-option.vue'
+import mdcSelect from './mdc-select.vue'
+import mdcOption from './mdc-option.vue'
 
 const components = {
-  'mdc-menu-select': mdcMenuSelect,
-  'mdc-menu-option': mdcMenuOption,
-  'mdc-native-select': mdcNativeSelect,
-  'mdc-native-option': mdcNativeOption,
-  'mdc-multi-select': mdcMultiSelect,
-  'mdc-multi-option': mdcMultiOption,
+  'mdc-select': mdcSelect,
+  'mdc-option': mdcOption,
 }
 
 export default {
@@ -43,6 +25,6 @@ export default {
       selectValues: [],
     }
   },
-  components 
+  components
 }
 </script>
